requirements.txt
iPython
Django==1.7
django-localflavor
MySQL-python
djangorestframework
markdown
django-filter
Pillow
south==0.8.4
circus
circus-web
chaussette
django-sslify 
django-jinja
rfc3339
django-selenium
selenium
importlib
gunicorn
gevent
psycopg2
django-websocket-redis
django-authhacks
mandrill
django-mandrill
django-http-proxy
pyvirtualdisplay